+++ napari/_qt/qt_viewer.py
@@ -16,8 +16,8 @@ from .qt_dims import QtDims
 from .qt_layerlist import QtLayerList
 from ..resources import resources_dir
 from ..util.theme import template
-from ..util.misc import (
-    str_to_rgb,
+from ..util.misc import str_to_rgb
+from ..util.interactions import (
     ReadOnlyWrapper,
     mouse_press_callbacks,
     mouse_move_callbacks,
