+++ napari/util/misc.py
@@ -4,10 +4,7 @@ from enum import Enum, EnumMeta
 import re
 import inspect
 import itertools
-from numpydoc.docscrape import FunctionDoc
 import numpy as np
-import wrapt
-import sys
 
 
 def str_to_rgb(arg):
@@ -49,41 +46,6 @@ def is_iterable(arg, color=False):
         return True
 
 
-def compute_max_shape(shapes, max_dims=None):
-    """Computes the maximum shape combination from the given shapes.
-
-    Parameters
-    ----------
-    shapes : iterable of tuple
-        Shapes to coombine.
-    max_dims : int, optional
-        Pre-computed maximum dimensions of the final shape.
-        If None, is computed on the fly.
-
-    Returns
-    -------
-    max_shape : tuple
-        Maximum shape combination.
-    """
-    shapes = tuple(shapes)
-
-    if max_dims is None:
-        max_dims = max(len(shape) for shape in shapes)
-
-    max_shape = [0] * max_dims
-
-    for dim in range(max_dims):
-        for shape in shapes:
-            try:
-                dim_len = shape[dim]
-            except IndexError:
-                pass
-            else:
-                if dim_len > max_shape[dim]:
-                    max_shape[dim] = dim_len
-    return tuple(max_shape)
-
-
 def formatdoc(obj):
     """Substitute globals and locals into an object's docstring."""
     frame = inspect.currentframe().f_back
@@ -94,32 +56,6 @@ def formatdoc(obj):
         return obj
     finally:
         del frame
-
-
-def segment_normal_vector(a, b):
-    """Determines the unit normal of the vector from a to b.
-
-    Parameters
-    ----------
-    a : np.ndarray
-        Length 2 array of first point
-    b : np.ndarray
-        Length 2 array of second point
-
-    Returns
-    -------
-    unit_norm : np.ndarray
-        Length the unit normal of the vector from a to b. If a == b,
-        then returns [0, 0]
-    """
-    d = b - a
-    normal = np.array([d[1], -d[0]])
-    norm = np.linalg.norm(normal)
-    if norm == 0:
-        unit_norm = np.array([0, 0])
-    else:
-        unit_norm = normal / norm
-    return unit_norm
 
 
 class StringEnumMeta(EnumMeta):
@@ -256,203 +192,3 @@ class CallSignature(inspect.Signature):
 
 
 callsignature = CallSignature.from_callable
-
-
-class ReadOnlyWrapper(wrapt.ObjectProxy):
-    """
-    Disable item and attribute setting with the exception of  ``__wrapped__``.
-    """
-
-    def __setattr__(self, name, val):
-        if name != '__wrapped__':
-            raise TypeError(f'cannot set attribute {name}')
-        super().__setattr__(name, val)
-
-    def __setitem__(self, name, val):
-        raise TypeError(f'cannot set item {name}')
-
-
-def mouse_press_callbacks(obj, event):
-    """Run mouse press callbacks on either layer or viewer object.
-
-    Note that drag callbacks should have the following form::
-        def hello_world(layer, event):
-            "dragging"
-            # on press
-            print('hello world!')
-            yield
-
-            # on move
-            while event.type == 'mouse_move':
-                print(event.pos)
-                yield
-
-            # on release
-            print('goodbye world ;(')
-
-    Parameters
-    ---------
-    obj : napari.components.ViewerModel or napar.layers.Layer
-        Layer or Viewer object to run callbacks on
-    event : Event
-        Mouse event
-    """
-    # iterate through drag callback functions
-    for mouse_drag_func in obj.mouse_drag_callbacks:
-        # exectute function to run press event code
-        gen = mouse_drag_func(obj, event)
-        # if function returns a generator then try to iterate it
-        if inspect.isgeneratorfunction(mouse_drag_func):
-            try:
-                next(gen)
-                # now store iterated genenerator
-                obj._mouse_drag_gen[mouse_drag_func] = gen
-                # and now store event that initially triggered the press
-                obj._persisted_mouse_event[gen] = event
-            except StopIteration:
-                pass
-
-
-def mouse_move_callbacks(obj, event):
-    """Run mouse move callbacks on either layer or viewer object.
-
-    Note that drag callbacks should have the following form::
-        def hello_world(layer, event):
-            "dragging"
-            # on press
-            print('hello world!')
-            yield
-
-            # on move
-            while event.type == 'mouse_move':
-                print(event.pos)
-                yield
-
-            # on release
-            print('goodbye world ;(')
-
-    Parameters
-    ---------
-    obj : napari.components.ViewerModel or napar.layers.Layer
-        Layer or Viewer object to run callbacks on
-    event : Event
-        Mouse event
-    """
-    if not event.is_dragging:
-        # if not dragging simply call the mouse move callbacks
-        for mouse_move_func in obj.mouse_move_callbacks:
-            mouse_move_func(obj, event)
-
-    # for each drag callback get the current generator
-    for func, gen in tuple(obj._mouse_drag_gen.items()):
-        # save the event current event
-        obj._persisted_mouse_event[gen].__wrapped__ = event
-        try:
-            # try to advance the generator
-            next(gen)
-        except StopIteration:
-            # If done deleted the generator and stored event
-            del obj._mouse_drag_gen[func]
-            del obj._persisted_mouse_event[gen]
-
-
-def mouse_release_callbacks(obj, event):
-    """Run mouse release callbacks on either layer or viewer object.
-
-    Note that drag callbacks should have the following form::
-        def hello_world(layer, event):
-            "dragging"
-            # on press
-            print('hello world!')
-            yield
-
-            # on move
-            while event.type == 'mouse_move':
-                print(event.pos)
-                yield
-
-            # on release
-            print('goodbye world ;(')
-
-    Parameters
-    ---------
-    obj : napari.components.ViewerModel or napar.layers.Layer
-        Layer or Viewer object to run callbacks on
-    event : Event
-        Mouse event
-    """
-    for func, gen in tuple(obj._mouse_drag_gen.items()):
-        obj._persisted_mouse_event[gen].__wrapped__ = event
-        try:
-            # Run last part of the function to trigger release event
-            next(gen)
-        except StopIteration:
-            pass
-        # Finally delete the generator and stored event
-        del obj._mouse_drag_gen[func]
-        del obj._persisted_mouse_event[gen]
-
-
-KEY_SYMBOLS = {
-    'Control': 'Ctrl',
-    'Shift': '⇧',
-    'Alt': 'Alt',
-    'Option': 'Opt',
-    'Meta': '⊞',
-    'Left': '←',
-    'Right': '→',
-    'Up': '↑',
-    'Down': '↓',
-    'Backspace': '⌫',
-    'Tab': '↹',
-    'Escape': 'Esc',
-    'Return': '⏎',
-    'Enter': '↵',
-}
-if sys.platform.startswith('darwin'):
-    KEY_SYMBOLS.update(
-        {'Control': '⌘', 'Alt': '⌥', 'Option': '⌥', 'Meta': '⌃'}
-    )
-elif sys.platform.startswith('linux'):
-    KEY_SYMBOLS.update({'Meta': 'Super'})
-
-
-def get_keybindings_summary(keymap, col='rgb(134, 142, 147)'):
-    """Get summary of keybindings in keymap.
-
-    Parameters
-    ---------
-    keymap : dict
-        Dictionary of keybindings.
-    col : str
-        Color string in format rgb(int, int, int) used for highlighting
-        keypress combination.
-
-    Returns
-    ---------
-    keybindings_str : str
-        String with summary of all keybindings and their functions.
-    """
-    keybindings_str = '<table border="0" width="100%">'
-    for key in keymap:
-        keycodes = [KEY_SYMBOLS.get(k, k) for k in key.split('-')]
-        keycodes = "+".join(
-            [f"<span style='color: {col}'><b>{k}</b></span>" for k in keycodes]
-        )
-        keybindings_str += (
-            "<tr><td width='80' style='text-align: right; padding: 4px;'>"
-            f"<span style='color: rgb(66, 72, 80)'>{keycodes}</span></td>"
-            "<td style='text-align: left; padding: 4px; color: #CCC;'>"
-            f"{get_function_summary(keymap[key])}</td></tr>"
-        )
-    keybindings_str += '</table>'
-    return keybindings_str
-
-
-def get_function_summary(func):
-    """Get summary of doc string of function."""
-    doc = FunctionDoc(func)
-    summary = ''
-    for s in doc['Summary']:
-        summary += s
-    return summary.rstrip('.')
